move batch 1 of ttnn tensor tests over
